+++ XSharpx.Examples/XSharpx.Examples/XSharpx.Examples.csproj
@@ -35,7 +35,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="PossibleExamples.cs" />
+    <Compile Include="AttemptExamples.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ItemGroup>
