+++ XSharpx.csproj
@@ -48,7 +48,7 @@
     <Compile Include="src\XSharpx\Tree.cs" />
     <Compile Include="src\XSharpx\Unit.cs" />
     <Compile Include="src\XSharpx\Iteratee.cs" />
-    <Compile Include="src\XSharpx\Possible.cs" />
+    <Compile Include="src\XSharpx\Attempt.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
